Backend/requirements.txt
flask
flask-cors
torch
torchvision
torchaudio
transformers
numpy
pandas
librosa
soundfile
opencv-python-headless
pillow
facenet-pytorch
scikit-learn
accelerate